requirements.txt
azure-core==1.30.2
azure-identity==1.17.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
idna==3.7
msal==1.29.0
msal-extensions==1.2.0
portalocker==2.10.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.2
